# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateEventRequest:
    properties:
      age_rating:
        type: string
      description:
        type: string
      director:
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      event_type:
        enum:
        - EVENT
        - MOVIE
        type: string
      language:
        type: string
      name:
        type: string
      release_date:
        type: string
    required:
    - duration_minutes
    - event_type
    - name
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  dto.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.UpdateEventRequest:
    properties:
      age_rating:
        type: string
      description:
        type: string
      director:
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      event_type:
        enum:
        - EVENT
        - MOVIE
        type: string
      language:
        type: string
      name:
        type: string
      release_date:
        type: string
    required:
    - duration_minutes
    - event_type
    - name
    type: object
info:
  contact: {}
  description: Event CRUD API for both normal events and movies.
  title: Event Service API
  version: "1.0"
paths:
  /events:
    get:
      description: List events with pagination and optional filters.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Event type
        in: query
        name: type
        type: string
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event or movie.
      parameters:
      - description: Create event request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create event
      tags:
      - events
  /events/{id}:
    delete:
      description: Soft-delete event by ID.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete event
      tags:
      - events
    get:
      description: Retrieve event detail by ID.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update event by ID.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Update event request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update event
      tags:
      - events
schemes:
- http
- https
swagger: "2.0"